stylesheet renaming fixes
