+++ src/frontend/src/stylesheets/pages/change-request-detail-page/proposed-solution-view.module.css
