+++ src/frontend/src/stylesheets/pages/change-request-detail-page/proposed-solutions-list.module.css
